requirements.txt
# Core dependencies
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0

# API clients
replicate>=0.25.0
openai>=1.0.0
stability-sdk>=0.8.0
google-genai>=1.0.0

# Image processing
opencv-python>=4.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0
